feat: like is now toggleLike and will switch between like and unlike based on if the user liked the post already
